requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
quart==0.19.3
quart_Ddscord==2.1.4